fix(agent-page): keep artifact tree alive across maximize remount
The agent right-pane renders `ArtifactPane` as two separate instances
(`Shell.Host` docked + `Shell.MaximizedOverlay`). Toggling maximize
unmounts one and mounts the other, and each mount re-ran the full O(N)
directory-tree materialization on the renderer main thread
(`useDirectoryTree` create + `rootFromSerialized` + `indexTree`, then
`projectArtifactTree`/`mergeLazyChildren`/`nodeById`/`filteredTree`) —
freezing the UI on every maximize/minimize. The main-side 500ms grace
window and `forceMount` did not cover this (renderer rematerialization;
two separate `Shell.Tabs` subtrees).

Lift the whole tree model into `useArtifactFileTreeModel`, owned by the
surviving `AgentRightPaneStateProvider` and exposed via a dedicated
context to the files panel only. `ArtifactPane` splits into a
presentational `ArtifactPaneView` (rendered by the agent path with the
lifted model) and a thin standalone wrapper that still owns its tree for
existing callers/tests. The tree is now created once and survives the
swap; the remounted view just reads an already-materialized, virtualized
list.
